Update _WIN32_WINNT to _WIN32_WINNT_VISTA
Also define WINVER used by some header files.
The change will help warn against use of API calls not in vista.
The original setting of this macro to  _NTDDI_WINXP was wrong and
permitted use of any API not just those supported by WinXP. Build
is unaffected as we currently do not have any function calls not
present in Vista.

Signed-off-by: Selva Nair <selva.nair@gmail.com>