Detect urls in echo message make them clickable
- Left clickng on http or https url will open it on the default browser

Several other URL schemes are detected and formatted as clickable
links, but we only support opening of http/https links.

Note on spaces in URLs: We unescape all %xx occurrences in the echo
message text so that %20 will be converted to space in plain text.
This means embedded spaces in URLs will not work even if written
as %20. An option is to use %2520 which will get conveted to %20
after the unescaping.
A better option is to enclose the URL in <>. If the
text inside <> starts with a valid scheme (http, https etc.),
the entire text including spaces is parsed as the URL.

Signed-off-by: Selva Nair <selva.nair@gmail.com>