Force script-security to SSEC_BUILT_IN by default
The setting is persisted in the registry as
HKCU\OpenVPN-GUI\config-name\script_security (DWORD)
with possible values 0 to 3 that has the same meaning
as described in  OpenVPN man page or -1 to mean do not
override the option (i.e, any value in the config file
will be effective).

The setting is reverted to default if the file changes
(tracked by a digest of the file). This feature is not
effective until the next commit where the digest function
is implemented.

The config file is locked against writes when a connection
is active. This ensures that if/when the file is re-read
by OpenVPN for a SIGHUP restart the contents have not
changed.

If no value is found in the registry the setting defaults
to 1 (SSEC_BUILT_IN).

TODO: add a dialog to edit the setting

Signed-off-by: Selva Nair <selva.nair@gmail.com>