Explicitly persist password-save preferences in registry
Currently user preference to save password is implicitly set
to true if a saved password is found in the registry. This makes
it difficult to toggle the save password flag from a dialog if
no previously saved password exists. Instead explicitly save
the preference.

This change helps implementation of the config-specific preferences
dialog (next commit). No user-visible changes.

Signed-off-by: Selva Nair <selva.nair@gmail.com>