Increase the buffer size for help message
The help message is truncated in some languages with
recent increases in textual information in the
usage string. Eg., see "openvpn-gui --help" with
Dutch selected as the language where the text
is about 2300 characters while the buffer is only 2048.

The window still fits on typical screen sizes, so
just increase the buffer size to 3000.

Signed-off-by: Selva Nair <selva.nair@gmail.com>