Use iservice even when gui is running under admin
We didn't use interactive service when gui was running
under admin because of some privilege escalation vulnerability in Vista.

Apparently this issue doesn't exist on Win7 and newer versions so
it is safe to use iservice on those systems.

Recent versions of mingw (like on Ubuntu18.04 which we use for
building installer) include versionhelpers.h so no need for compat layer.

Signed-off-by: Lev Stipakov <lev@openvpn.net>