Allow --command action args to work also as --action args
When sending multiple commands to a running instance of the GUI,
simplify usage like

openvpn-gui.exe --command connect vpn1 --command connect vpn2
      --command disconnect vpn3

by omitting --command so that one could write

openvpn-gui.exe --connect vpn1 --connect vpn2 --disconnect vpn3

The form "--command action .. " was introduced originally to indicate that
these actions are only understood if there is an instance of the gui
already running as it just triggers a message for that action to be sent to
the first instance. But it becomes tedious when multiple commands are to
be specified.

This patch makes --action arg  to also work the same way as
--command action arg. A side effect is that some of such options do not
make sense when launching the first instance of the GUI. In such cases,
the option is ignored but an error is logged to the eventlog. In particular,
use of options such as --command disconnect or --disconnect with the first
instance are no longer treated as a fatal error.

The allowed values of action are unchanged:
  connect
  disconnect
  reconnect
  disconnect_all
  exit
  status
  silent_connection
  import
  rescan
out of which connect, silent_connection and import are the only
one's that could be handled by the first instance.

Example:

openvpn-gui.exe --connect vpn1 --connect vpn2

will start vpn1 and vpn2 if OpenVPN-GUI is not already running,
or cause a command to start those sent to a running instance.

Signed-off-by: Selva Nair <selva.nair@gmail.com>