Save active connections on exit and auto-start in next session
New feature: any connection that is not disconnected on exit
will auto-connect when the GUi is started the next time.

There is no option to toggle auto-connect of any profile. Instead,
just connect normally and leave the connection open while closing the
GUI directly or indirectly (on logout, for example). Such a connection
will auto-connect when the GUI is started the next time.

If auto-connect is not desired for a particular connection, stop it
before exit. Or, the whole feature may be disabled in the setings menu
(implemented in next commit).

Signed-off-by: Selva Nair <selva.nair@gmail.com>