fixup! Check return value of SetProp
Remove CHECK_NULL_PARAM and other over-zealous checks on the
return value of GetProp.

In pkcs11.c if SetProp() fails just do not use bold font for
header instead of leaking the font resource.
Also correct a bad fixup in commit 80697ecae6: hfontProp was not set!

(to be squashed with previous commit)

Signed-off-by: Selva Nair <selva.nair@gmail.com>