Fix memory leak in PLAP
When LogonUI unloads and reloads the PLAP dll, or when the provider
is released and re-created, memory allocated for config list and
groups leak.

- Fix by freeing config list and groups in DeleteUI

We do not call this before exiting WinMain in the GUI code,
as its hard to do it safely -- have to ensure all status
threads have terminated. Anyway, freeing is only cosmetic in this case.

Signed-off-by: Selva Nair <selva.nair@gmail.com>