PLAP: support for webauth
Instead of opening the browser window in PLAP context (which is insecure),
display a dialog with QR code and ask user to scan it on mobile device.

QR code generation is taken from MIT-licensed library:

  https://github.com/nayuki/QR-Code-generator/blob/master/c/qrcodegen.c

Before I tried vcpkg libqrencode package but gave up after
10 mins of dependencies building (it requires msys2 etc).

Note that you need to add

  setenv IV_SSO webauth

to PLAP profile to make it work.

Fixes: https://github.com/OpenVPN/openvpn-gui/issues/687

Signed-off-by: Lev Stipakov <lev@openvpn.net>