Avoid blocking calls during WM_OVPN_ECHOMSG processing
We send this message from the connection thread to the main
thread using a blocking SendMesssage as the echo message to display is
cleared soon after. This sometimes leads to timeouts (and
thus long waits). Avoid timeout using Async ShowWindow() and delaying
SetForegroundWindow() as these could otherwise wait if the current
foreground Window is controlled by the calling thread.

Signed-off-by: Selva Nair <selva.nair@gmail.com>